program is able to write dictionary to file updated Dictionary.h and Dictionary.cpp to change valType from int to double

+++ Dictionary.h
@@ -19,7 +19,7 @@ pa8
 
 // Exported types -------------------------------------------------------------
 typedef std::string keyType;  // must be comparable using <, <=, >, >=, ==, !=
-typedef int         valType;
+typedef double         valType;
 
 class Dictionary{
